Fixes tests on PHP 8.0
