+++ tests/Routing/Enums.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Illuminate\Tests\Routing;
+
+enum CategoryEnum
+{
+    case People;
+    case Fruits;
+}
+
+enum CategoryBackedEnum: string
+{
+    case People = 'people';
+    case Fruits = 'fruits';
+}
